requirements.txt
# pip install -r requirements.txt

PyQt5
torchtext
matplotlib

konlpy
# need to downgrade tweepy
tweepy==3.7